+++ src/components/About/AboutCard.tsx
