+++ src/components/Anime/AnimeCard.tsx
