+++ src/components/Anime/AnimeList.tsx
