+++ src/components/Anime/AnimeListItem.tsx
