+++ src/components/Title/index.tsx
@@ -1,7 +1,7 @@
 import Head from "next/head";
 import React from "react";
 import type { FC } from "../../lib/types";
-import config from "../../config.json";
+import config from "../../../config.json";
 
 const Title: FC<{ title: string }> = ({ title }) => {
 	return (
