+++ src/styles/components/about.scss
