+++ src/styles/components/anime.scss
