+++ src/styles/components/discord.scss
