+++ src/styles/components/modal.scss
