+++ src/styles/pages/404.scss
