+++ src/styles/pages/about.scss
