+++ src/styles/pages/index.scss
