+++ src/styles/pages/stats.scss
